requirements.txt
boto3
aws-lambda-powertools
python-jose[cryptography]